Final Update

+++ application/controllers/JenisFaskes.php
@@ -58,6 +58,9 @@ class JenisFaskes extends CI_Controller
 	}
     public function delete($id)
     {
+        if($this->session->userdata('role') != "administrator"){
+			redirect(base_url("auth"));
+		}
         $this->load->model('Jenis_faskes_model', 'faskes');
 		$data['id'] = $id;
 		$this->faskes->delete($data);
